Shitload of stuff, Phase 2 initial implementation
The AI seems stable, the lasershot is good, the spinshot is lacking the
animation at this point.

A few things:
- I added an "invincible" property to enemy_PAR. The boss (and orbs) now
are invincible until they start doing stuff.
- We seriously need a projectile_PAR for friendly.
- The code in the collision between projectiles and enemies also handles
enemy death. WTF? Fix this.
- Projectiles play their sound on spawn. This, it turns out, is really
BAD when you spawn a shitload of them all at once. Move the audio code
to the weapon/enemy that spawns it.